==> code/2. select_cities_ninja_api.ipynb <==
### Why select cities?
The next step (code file #2) will undertake webscraping to extract population and geographical data on selected countries. While it suffices to pick cities of our choice, I will take an automatized approach to city selection. The following two steps are undertaken:
1. I select a list of countries that Gans is present.
2. Then I use Ninja api to pick large cities in these countries with population more than 1,000,000. 

#### Import packages

import pandas as pd
import requests

# Get the api key
from keys import Ninja_API_Key

#### Select countries

countries=["DE", "FR", "IT", "PL", "PT", "ES", "GB", "CH", "NL", "BE"]

#### Pick up the most populated cities from selected countries

# Define the sections that will together form the url.
url = "https://api.api-ninjas.com/v1/city"
header={"X-Api-Key": Ninja_API_Key}

# Define the sections that will together form the url.
cities_json={}
for i in countries:
    # Reference the sections in the request.
    cities = requests.request("GET", url, headers=header, params={"country": i, "min_population": "1000000", "limit": "5"})
    city_json = cities.json()
    cities_json[i]=city_json
    cities_json

print(cities_json.keys())
print(cities_json)
print(cities_json.values())

#### Pick up columns of interest

cities_df = pd.concat([pd.DataFrame(cities, columns=['name', 'country', 'latitude', 'longitude', 'is_capital']) for cities in cities_json.values()], ignore_index=True)
cities_df

#### Rename columns

cities_df_cleaned = cities_df.rename(columns={"country": "country_2c", "name": "City"})
cities_df_cleaned

#### Save the df as a csv file

cities_df_cleaned.to_csv("data/cities_df_cleaned.csv", sep=',', index=False, encoding='utf-8')
