==> code/4. SQL_integration_sqlalchemy.ipynb <==
### Create a unique identifier for city column using SQL
In this exercise, I will perform the following steps:
- Establish a connection link with the SQL DB in MySQL workbench
- Move the cities dataframe (columns: "City" and "Elevation") to SQL DB uniquely identifying the city names (table column "city_id")
- Read the SQL table in python and merge city_id back into the webscraped dataframe
- Move data to the following SQL tables:
    1. Countries (with columns: "city_id", "Country", "country_2c")
    2. Populations (with columns: "city_id", "Population")
    3. Coordinates (with columns: "city_id", "latitude", "longitude")

#### Import packages

import pandas as pd
import sqlalchemy

# Get the api key
from keys import MySQL_bootcamp

#### Create the connection link

schema = "gans_cities"
host = "127.0.0.1"
user = "root"
password = MySQL_bootcamp
port = 3306

connection_string = f'mysql+pymysql://{user}:{password}@{host}:{port}/{schema}'

#### Import the merged dataframe

df_cities_ws_cleaned = pd.read_csv("data/df_cities_ws_cleaned.csv")

#### Create a unique identifier for the city using SQL

df = df_cities_ws_cleaned.copy()
df_city = df[["City", "Elevation (in m)"]]
df_city=df_city.rename(columns= {"Elevation (in m)": "Elevation"})

df_city.to_sql('cities',
                  if_exists='append',
                  con=connection_string,
                  index=False)

#### Import SQL table and merge city_id in the original dataframe

df_city_sql = pd.read_sql("cities", con=connection_string)
df_sql_city_id = df.merge(df_city_sql, on = "City", how="left")
df_sql_city_id

#### Create countries table in SQL

df_country = df_sql_city_id[["city_id", "country_2c", "Country"]]

df_country.to_sql('countries',
                  if_exists='append',
                  con=connection_string,
                  index=False)

#### Create a population table in SQL

df_pop = df_sql_city_id[["city_id", "Population"]]
df_pop = df_pop.copy()
df_pop["Year_retrieved"] = 2024

df_pop.to_sql('populations',
                  if_exists='append',
                  con=connection_string,
                  index=False)

#### Create a coordinates table

df_coord = df_sql_city_id[["city_id", "latitude", "longitude"]]
df_coord.to_sql('coordinates',
                  if_exists='append',
                  con=connection_string,
                  index=False)

#### Save dataframe as csv

df_sql_city_id.to_csv("data/df_sql_city_id.csv", sep=',', index=False, encoding='utf-8')
