==> code/5. weather_data_OW_api.ipynb <==
### OpenWeather api to collect weather data
Now I will collect weather forecast data for the selected cities from the OpenWeather api ([API Documentation](https://openweathermap.org/forecast5)). The api allows us a free access to **5 Day / 3 Hour Forecast**: a 5 day forecast data with 3-hour step. 


Relevant for the Gans use case, I will collect information on the following:
- *'Temp'* - Average temperature
- *'Feels_like'* - Avg. temp felt. 
- *'Sunrise'* - Datetime for the sunrise 
- *'Sunset'* - Datetime for the sunset 
- *'Weather'* - Overall description of the weather (Rain, Snow, Clouds etc.)
- *'Snow'* - Avg. snow
- *'Rain'* - Avg. rain
- *'Wind'* -  Avg. wind speed

import pandas as pd
import requests
import sqlalchemy
from datetime import datetime
from pytz import timezone

# Get the api keys
from keys import MySQL_bootcamp, OW_API_Key

#### Read cleaned webscraped city_df

df_sql_city_id = pd.read_csv("data/df_sql_city_id.csv")
df_sql_city_id

#### Open weather map data collection

API_key = OW_API_Key 
city_weather = df_sql_city_id["City"].unique()
cities_weather_json = []

berlin_timezone = timezone('Europe/Berlin')

def cities_dataframe(cities):
    cities_data = []
    for i in df_sql_city_id["City"]:
        cities_weather = requests.get(f"http://api.openweathermap.org/data/2.5/forecast?q={i}&appid={API_key}&units=metric")
        cities_weather_json = cities_weather.json()
        # As we are now using the data from our relational database
        # the city should reflect the city_id and not the city name
        city_id = df_sql_city_id.loc[df_sql_city_id["City"] == i, "city_id"].values[0]
        retrieval_time = datetime.now(berlin_timezone).strftime("%Y-%m-%d %H:%M:%S")
        
        for entry in cities_weather_json["list"]:
            city_data_df = {}
            city_data_df["city_id"]= city_id
            # city_data_df["City"] = cities_weather_json["city"]["name"]
            city_data_df["Datetime"] = entry['dt_txt'] if 'dt_txt' in entry else 0 # Check if 'rain' key exists in the entry, set to 0 if not present
            city_data_df["Temp"] = entry["main"]["temp"]
            city_data_df["Feels_like"] = entry["main"]["feels_like"]
            # city_data_df["Coordinates"] = cities_weather_json["city"]["coord"]
            city_data_df["Sunrise"] = pd.to_datetime(cities_weather_json["city"]["sunrise"], unit='s')
            city_data_df["Sunset"] = pd.to_datetime(cities_weather_json["city"]["sunset"], unit='s')
            # city_data_df["Population"] = cities_weather_json["city"]["population"]
            city_data_df["Weather"] = entry['weather'][0]['description'] 
            city_data_df["Snow"] = entry['snow']['3h'] if 'snow' in entry and '3h' in entry['snow'] else 0 # Check if 'rain' key exists in the entry, set to 0 if not present
            city_data_df["Rain"] = entry['rain']['3h'] if 'rain' in entry and '3h' in entry['rain'] else 0 # Check if 'rain' key exists in the entry, set to 0 if not present            
            city_data_df["Wind"] = entry['wind']['speed'] if 'wind' in entry and 'speed' in entry['wind'] else 0 # Check if 'rain' key exists in the entry, set to 0 if not present
            city_data_df["data_retrieved_at"]= retrieval_time
            cities_data.append(city_data_df.copy())  # Make sure to copy the dictionary to avoid overwriting
                 
    return pd.DataFrame(cities_data)

cities_weather_df = cities_dataframe(city_weather).sort_values(by=["city_id", "Datetime"], ascending=[True, True])
cities_weather_df.head()

cities_weather_df.info()

#### Merge with city_df

df_city_id_weather = cities_weather_df.merge(df_sql_city_id, how="outer", on="city_id")

#### Save df as a csv

df_city_id_weather.to_csv("data/df_city_id_weather.csv", sep=',', index=False, encoding='utf-8')

#### Create a weather table in SQL DB

# Create a connecting link
schema = "gans_cities"
host = "127.0.0.1"
user = "root"
password = MySQL_bootcamp
port = 3306

connection_string = f'mysql+pymysql://{user}:{password}@{host}:{port}/{schema}'

Weather_df = df_city_id_weather[['city_id', 'Datetime', 'Temp', 'Feels_like', 'Sunrise', 'Sunset', 'Weather', 'Snow', 'Rain', 'Wind']]
Weather_df.to_sql('weather',
                  if_exists='append',
                  con=connection_string,
                  index=False)
