==> code/6. flights_data_aerodatabox_api.ipynb <==
### Getting flights data for the cities
Gans wants to know when the flights arrive to the city of interest to be able to predict the escooter usage. To obtains flights arrival times for the day of interest, I use the AeroDataBox api ([API Documentation](https://rapidapi.com/aedbx-aedbx/api/aerodatabox/details)). This code will generate two SQL tables:

1. **Airports icao codes**: A function *icao_airport_codes(latitudes, longitudes)* will take in city coordinates and find airports closer to the selected cities. Importantly, it will store city icao codes that will be used to obtain the flights data.  
2. **Flights data**: A function *get_arrival_data(icao_list)* will take in airport ICAO codes and produce the following flights data columns. Notably, the flight data for the next day will be produced to be able to forecast e-scooter traffic. 
- *"arrival_airport_icao"* - Arrival airport's ICAO code
- "departure_airport_icao"* - Departure airport's ICAO code
- *"departure_airport_name"* - Departure airport's name
- *"scheduled_arrival_time"* - Scheduled arrival time
- *"flight_number"* - Flight number
- *"data_retrieved_at"* - Date of the data retrieval

#### Import libraries and cities df

import pandas as pd
import requests
import sqlalchemy

import datetime
from pytz import timezone
from datetime import datetime, timedelta

import warnings
warnings.filterwarnings('ignore')

# Get the api key
from keys import AeroDatabox, MySQL_bootcamp

df_cities = pd.read_csv("data/df_sql_city_id.csv")
df_cities

#### Get airport icao codes for selected cities

# Creating a function to get airports for the city
def icao_airport_codes(latitudes, longitudes):
  #assert len(latitudes) == len(longitudes)
  list_for_df = []
  for i in range(len(latitudes)):
    url = "https://aerodatabox.p.rapidapi.com/airports/search/location"
    querystring = {"lat":latitudes[i],"lon":longitudes[i],"radiusKm":"50","limit":"5","withFlightInfoOnly":"true"}
    headers = {
      "X-RapidAPI-Host": "aerodatabox.p.rapidapi.com",
      "X-RapidAPI-Key": AeroDatabox
    }
    response = requests.request("GET", url, headers=headers, params=querystring)
    list_for_df.append(pd.json_normalize(response.json()['items']))
  return pd.concat(list_for_df, ignore_index=True)

df_airports = pd.DataFrame(icao_airport_codes(df_cities["latitude"], df_cities["longitude"]))
df_airports

Restrict the dataframe to preferred information

df_airports = df_airports[["icao", "municipalityName"]]

#### Merge airports dataframe with cities df
First let's rename cities to match city_df city names

df_airports.loc[df_airports["municipalityName"]=="Newcastle upon Tyne", "municipalityName"]="Newcastle"
df_airports.loc[df_airports["municipalityName"]=="Seville", "municipalityName"]="Sevilla"
df_airports.iloc[27, df_airports.columns.get_loc('municipalityName')] = "The Hague"

Merge two dataframes and restrict the merged df to contains needed columns

df_airports_merged = df_cities.merge(df_airports, how="left", left_on="City", right_on="municipalityName")
df_airports_merged = df_airports_merged[["city_id", "icao"]]
df_airports_merged

#### Create airports table in SQL DB

# Create a connecting link
schema = "gans_cities"
host = "127.0.0.1"
user = "root"
password = MySQL_bootcamp
port = 3306

connection_string = f'mysql+pymysql://{user}:{password}@{host}:{port}/{schema}'


df_airports_merged.to_sql('airports',
                  if_exists='append',
                  con=connection_string,
                  index=False)

#### Save the airports df

df_airports_merged.to_csv("data/df_airports_merged.csv", sep=',', index=False, encoding='utf-8')

#### Get flights data

# Create a function to get arrival info at the selected airport
def get_arrival_data(icao_list):
  api_key = AeroDatabox

  berlin_timezone = timezone('Europe/Berlin')
  today = datetime.now(berlin_timezone).date()
  tomorrow = (today + timedelta(days=1))

  flight_items = []

  for icao in icao_list:
    # the api can only make 12 hour calls, therefore, 2 12 hour calls make a full day
    # using the nested lists below we can make a morning call and extract the data
    # then make an afternoon call and extract the data
    times = [["00:00","11:59"],
             ["12:00","23:59"]]

    for time in times:
      url = f"https://aerodatabox.p.rapidapi.com/flights/airports/icao/{icao}/{tomorrow}T{time[0]}/{tomorrow}T{time[1]}"

      querystring = {"withLeg":"true",
                    "direction":"Arrival",
                    "withCancelled":"false",
                    "withCodeshared":"true",
                    "withCargo":"false",
                    "withPrivate":"false"}

      headers = {
          'x-rapidapi-host': "aerodatabox.p.rapidapi.com",
          'x-rapidapi-key': api_key
          }

      response = requests.get(url, headers=headers, params=querystring)

      flights_json = response.json()

      retrieval_time = datetime.now(berlin_timezone).strftime("%Y-%m-%d %H:%M:%S")

      for item in flights_json["arrivals"]:
        flight_item = {
            "arrival_airport_icao": icao,
            "departure_airport_icao": item["departure"]["airport"].get("icao", None),
            "departure_airport_name": item["departure"]["airport"].get("name", None),
            "scheduled_arrival_time": item["arrival"]["scheduledTime"].get("local", None),
            "flight_number": item.get("number", None),
            "data_retrieved_at": retrieval_time
        }

        flight_items.append(flight_item)

  flights_df = pd.DataFrame(flight_items)
  flights_df["scheduled_arrival_time"] = flights_df["scheduled_arrival_time"].str[:-6]
  flights_df["scheduled_arrival_time"] = pd.to_datetime(flights_df["scheduled_arrival_time"])
  flights_df["data_retrieved_at"] = pd.to_datetime(flights_df["data_retrieved_at"])

  return flights_df

#### Get tomorrow's flights arrival data for London Heathrow airport

df_airport_london = df_airports.loc[df_airports["icao"]=="EGLL", "icao"]
# Specify the airport code ("EGLL") as a list
airport_codes = list(df_airport_london.values)
flights_df = get_arrival_data(airport_codes)
flights_df

flights_df["scheduled_arrival_time"].min()

flights_df["scheduled_arrival_time"].max()

#### Read airports table from SQL DB to get airport_id

df_airports_sql = pd.read_sql("airports", con=connection_string)
df_airports_sql=df_airports_sql[['airport_id', 'icao']]
df_airports_sql

#### Merge airports table from SQL DB to get airport_id

flights_merged_df= flights_df.merge(df_airports_sql, how="left", left_on="arrival_airport_icao", right_on="icao")
flights_merged_df.columns

flights_merged_df = flights_merged_df.drop("icao", axis=1)

#### Save the flights df

flights_merged_df.to_csv("data/df_flight_arrivals.csv", sep=',', index=False, encoding='utf-8')

#### Create a flights table in SQL DB

flights_merged_df.to_sql('flights',
                  if_exists='append',
                  con=connection_string,
                  index=False)
